Move SchemaMetaDataPersistService to infra module
