+++ shardingsphere-governance/shardingsphere-governance-core/src/test/java/org/apache/shardingsphere/governance/core/registry/cache/subscriber/ScalingRegistrySubscriberTest.java
@@ -70,6 +70,6 @@ public final class ScalingRegistrySubscriberTest {
 //    @SneakyThrows({IOException.class, URISyntaxException.class})
 //    private String readYAML() {
 //        return Files.readAllLines(Paths.get(ClassLoader.getSystemResource("yaml/regcenter/data-schema-rule.yaml").toURI()))
-//                .stream().filter(each -> !each.startsWith("#")).map(each -> each + System.lineSeparator()).collect(Collectors.joining());
+//                .stream().map(each -> each + System.lineSeparator()).collect(Collectors.joining());
 //    }
 }
