+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/config/persist/ConfigCenterTest.java
@@ -56,8 +56,6 @@ public final class ConfigCenterTest {
     
     private static final String SCHEMA_RULE_YAML = "yaml/configcenter/data-schema-rule.yaml";
     
-    private static final String GLOBAL_RULE_YAML = "yaml/configcenter/data-global-rule.yaml";
-    
     @Mock
     private DataSourcePersistService dataSourceService;
     
@@ -139,7 +137,6 @@ public final class ConfigCenterTest {
     
     @SneakyThrows({IOException.class, URISyntaxException.class})
     private String readYAML(final String yamlFile) {
-        return Files.readAllLines(Paths.get(ClassLoader.getSystemResource(yamlFile).toURI()))
-                .stream().filter(each -> !each.startsWith("#")).map(each -> each + System.lineSeparator()).collect(Collectors.joining());
+        return Files.readAllLines(Paths.get(ClassLoader.getSystemResource(yamlFile).toURI())).stream().map(each -> each + System.lineSeparator()).collect(Collectors.joining());
     }
 }
